requirements/base.txt
fastapi
uvicorn
python-decouple
dnspython
pymongo
pydantic
motor
pymongo[srv]
python-jose
uvicorn
python-decouple
dnspython
pymongo
pydantic
motor
pymongo[srv]
python-jose
passlib
sendgrid
